refactor: melhora a estrutura do todo.controller.ts
